bugs fixed. The merging of Robin's work and mine has done well.
