+++ src/main/java/dci/j24e01/OurShop/controllers/AppController.java
@@ -15,12 +15,5 @@ public class AppController {
         return "index";
     }
 
-    @ModelAttribute("navLinks")
-    public List<NavLink> getNavLinks(){
-        return List.of(
-                new NavLink("Home", "/"),
-                new NavLink("Categories", "/categories"),
-                new NavLink("Products", "/products")
-        );
-    }
+
 }
